refactor: split the create invoice page into multiple components
